% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_location.R
\name{get_location}
\alias{get_location}
\title{get_location}
\usage{
get_location(token = NA, only_count = FALSE, ...)
}
\arguments{
\item{token}{Character string api-key provided by RMPC.}

\item{only_count}{Boolean. Returns record count}

\item{...}{Any RMIS location table field name (e.g. reporting_agency, psc_basin, location_type, etc.)}
}
\description{
This function pulls location data from the RMIS database.
Any fields in the location table are passible through this function.
Please review data documentation and understand data documentation before using this package.
}
\examples{
## get type 2  locations reported by ADFG
\dontrun{
adfg2<-get_locaton(token="your-api-key", reporting_agency="ADFG", location_type=2)
}

}
