% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recovery.R
\name{get_recovery}
\alias{get_recovery}
\title{get_recovery}
\usage{
get_recovery(token = NA, only_count = FALSE, ...)
}
\arguments{
\item{token}{Character string api-key provided by RMPC.}

\item{only_count}{Boolean. Returns record count}

\item{...}{Any RMIS recovery table field name (e.g. reporting_agency, species, run_year, etc.)}
}
\description{
This function pulls recovery data from the RMIS database.
Any fields in the recovery table are passible through this function.
Please review data documentation and understand data documentation before using this package.
Queries > 500,000 records are not allowed.
Querying 500k records would take a long time so ideally queries will be substantially smaller than that.
If you reeeeaaaally want all 10 million + records you can loop this function by run year.
}
\examples{
## get chinook recovery for 1990 reported by ADFG of species 1
\dontrun{
adfg1990<-get_recovery(token="your-api-key", reporting_agency="ADFG", run_year=1990, species = 1)
}
}
