% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mget_release.R
\name{mget_release}
\alias{mget_release}
\title{mget_release}
\usage{
mget_release(token = api_key, only_count = FALSE, ...)
}
\arguments{
\item{token}{Character string api-key provided by RMPC.}

\item{only_count}{Boolean. Returns record count}

\item{...}{Any RMIS release table field name (e.g. reporting_agency, species,
brood_year, etc.). Single values or vectors may be passed.}
}
\description{
This function is a wrapper for the get_release function that
allows for pulling data from RMIS over a range of parameter values.
Any fields in the release table are passable with either a single value
or vector of values. Please review data documentation and understand data
documentation before using this package.
}
